Clean up user test factories
